- updat type

+++ database/lvdb/customtoken.go
@@ -85,9 +85,9 @@ func (db *db) StoreCustomTokenPaymentAddresstHistory(tokenID *common.Hash, tx *t
 			return err
 		}
 		// old value: {value}-unspent-unreward/reward
-		values := strings.Split(string(value),string(spliter))
+		values := strings.Split(string(value), string(spliter))
 		// new value: {value}-spent-unreward/reward
-		newValues := values[0] + string(spliter) + string(spent) + string(spliter) +  values[2]
+		newValues := values[0] + string(spliter) + string(spent) + string(spliter) + values[2]
 		if err := db.lvdb.Put([]byte(paymentAddressKey), []byte(newValues), nil); err != nil {
 			return err
 		}
@@ -116,7 +116,7 @@ func (db *db) StoreCustomTokenPaymentAddresstHistory(tokenID *common.Hash, tx *t
 /*
 	Return a list of all address with balance > 0
 */
-func (db *db) GetCustomTokenListPaymentAddress(tokenID *common.Hash) ([][]byte, error){
+func (db *db) GetCustomTokenListPaymentAddress(tokenID *common.Hash) ([][]byte, error) {
 	results := [][]byte{}
 	//tempResults := make(map[string]int)
 	tempsResult := make(map[string]bool)
@@ -127,7 +127,7 @@ func (db *db) GetCustomTokenListPaymentAddress(tokenID *common.Hash) ([][]byte, 
 		value := string(iter.Value())
 		keys := strings.Split(key, string(spliter))
 		values := strings.Split(value, string(value))
-		if strings.Compare(values[1],string(unspent)) == 0 {
+		if strings.Compare(values[1], string(unspent)) == 0 {
 			// Uncomment this to get balance of all account
 			//i, ok := tempResults[keys[3]]
 			//if ok == false {
@@ -152,8 +152,8 @@ func (db *db) GetCustomTokenListPaymentAddress(tokenID *common.Hash) ([][]byte, 
 /*
 	Return a list of all address with balance > 0
 */
-func (db *db) GetCustomTokenListPaymentAddressesBalance(tokenID *common.Hash) (map[client.PaymentAddress]int, error){
-	results := make(map[client.PaymentAddress]int)
+func (db *db) GetCustomTokenListPaymentAddressesBalance(tokenID *common.Hash) (map[client.PaymentAddress]uint64, error) {
+	results := make(map[client.PaymentAddress]uint64)
 	//tempsResult := make(map[string]bool)
 	prefix := string(tokenPaymentAddressPrefix) + tokenID.String()
 	iter := db.lvdb.NewIterator(util.BytesPrefix([]byte(prefix)), nil)
@@ -162,7 +162,7 @@ func (db *db) GetCustomTokenListPaymentAddressesBalance(tokenID *common.Hash) (m
 		value := string(iter.Value())
 		keys := strings.Split(key, string(spliter))
 		values := strings.Split(value, string(value))
-		if strings.Compare(values[1],string(unspent)) == 0 {
+		if strings.Compare(values[1], string(unspent)) == 0 {
 			// Uncomment this to get balance of all account
 			paymentAddress := client.PaymentAddress{}
 			paymentAddress.FromBytes([]byte(keys[3]))
@@ -170,8 +170,8 @@ func (db *db) GetCustomTokenListPaymentAddressesBalance(tokenID *common.Hash) (m
 			if ok == false {
 				fmt.Println("ERROR geting value in GetCustomTokenAccountHistory of account", paymentAddress)
 			}
-			balance,_ := strconv.Atoi(values[0])
-			i += balance
+			balance, _ := strconv.Atoi(values[0])
+			i += uint64(balance)
 			results[paymentAddress] = i
 
 		}
@@ -179,13 +179,14 @@ func (db *db) GetCustomTokenListPaymentAddressesBalance(tokenID *common.Hash) (m
 	iter.Release()
 	return results, nil
 }
+
 /*
 	Get a list of UTXO that can be reward all payment address
 	key: payment address
 	value: a list of utxo
 	Each utxo consist of two part: first 32 byte is txHash, the rest is voutIndex
 */
-func (db *db) GetCustomTokenListUnrewardUTXO(tokenID *common.Hash) (map[client.PaymentAddress][][]byte, error){
+func (db *db) GetCustomTokenListUnrewardUTXO(tokenID *common.Hash) (map[client.PaymentAddress][][]byte, error) {
 
 	results := make(map[client.PaymentAddress][][]byte)
 	prefix := string(tokenPaymentAddressPrefix) + tokenID.String()
@@ -196,22 +197,23 @@ func (db *db) GetCustomTokenListUnrewardUTXO(tokenID *common.Hash) (map[client.P
 		keys := strings.Split(key, string(spliter))
 		values := strings.Split(value, string(value))
 		// get unspent and unreward transaction output
-		if (strings.Compare(values[1],string(unspent)) == 0) && (strings.Compare(values[2],string(unreward)) == 0) {
+		if (strings.Compare(values[1], string(unspent)) == 0) && (strings.Compare(values[2], string(unreward)) == 0) {
 			paymentAddress := client.PaymentAddress{}
 			paymentAddress.FromBytes([]byte(keys[3]))
 			//utxo := keys[4] + string(spliter) + keys[5]
-			utxo := append([]byte(keys[4]), []byte(keys[5])[:]... )
+			utxo := append([]byte(keys[4]), []byte(keys[5])[:]...)
 			results[paymentAddress] = append(results[paymentAddress], utxo)
 		}
 	}
 	iter.Release()
 	return results, nil
 }
+
 /*
 	Get a list of UTXO of one address
 	Return a list of UTXO, each UTXO has format: first 32 byte is txHash, the rest is voutIndex
 */
-func (db *db) GetCustomTokenPaymentAddressUTXO(tokenID *common.Hash, paymentAddress client.PaymentAddress) ([][]byte, error){
+func (db *db) GetCustomTokenPaymentAddressUTXO(tokenID *common.Hash, paymentAddress client.PaymentAddress) ([][]byte, error) {
 	prefix := string(tokenPaymentAddressPrefix) + tokenID.String() + string(spliter) + string(paymentAddress.ToBytes())
 	results := [][]byte{}
 	iter := db.lvdb.NewIterator(util.BytesPrefix([]byte(prefix)), nil)
@@ -221,19 +223,20 @@ func (db *db) GetCustomTokenPaymentAddressUTXO(tokenID *common.Hash, paymentAddr
 		keys := strings.Split(key, string(spliter))
 		values := strings.Split(value, string(value))
 		// get unspent and unreward transaction output
-		if (strings.Compare(values[1],string(unspent)) == 0) {
+		if (strings.Compare(values[1], string(unspent)) == 0) {
 			//utxo := keys[4] + string(spliter) + keys[5]
-			utxo := append([]byte(keys[4]), []byte(keys[5])[:]... )
+			utxo := append([]byte(keys[4]), []byte(keys[5])[:]...)
 			results = append(results, utxo)
 		}
 	}
 	iter.Release()
 	return results, nil
 }
+
 /*
 	Update UTXO from unreward -> reward
 */
-func (db *db) UpdateRewardAccountUTXO(tokenID *common.Hash, paymentAddress client.PaymentAddress, txHash *common.Hash, voutIndex int) (error){
+func (db *db) UpdateRewardAccountUTXO(tokenID *common.Hash, paymentAddress client.PaymentAddress, txHash *common.Hash, voutIndex int) (error) {
 	key := string(tokenPaymentAddressPrefix) + tokenID.String() + string(spliter) + string(paymentAddress.ToBytes()) + string(spliter) + txHash.String() + string(spliter) + strconv.Itoa(voutIndex)
 	_, err := db.hasValue([]byte(key))
 	if err != nil {
